refactor(search): delete search::solve shim, fix lib.rs facade drift
The type split left search::solve as a re-export shim whose only real
contents were CandidateAttempt/MultiSolveResult. Relocate those into
search::result, delete the shim, repoint internal imports and the one
python consumer at canonical search::options/search::result paths, and
fix the lib.rs facade which still re-exported via the stale solve::* path.
