+++ Ion.lock
@@ -1,102 +1,118 @@
 [[skill]]
 name = "brainstorming"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/brainstorming"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:5a4ac2fb5a057c07d7d4f362a8385de541a1ec0c4d3ebbd0028a873b9519c933"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:1518efecf35139a9a5b9e60906b71de874af08322200fab18a54848cea8c15fe"
 
 [[skill]]
 name = "dispatching-parallel-agents"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/dispatching-parallel-agents"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
 checksum = "sha256:e21044ae7496fc285f1df9ff3fb518a3a20bea6f5216c1f59f4987e12c769a0a"
 
 [[skill]]
 name = "executing-plans"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/executing-plans"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:2f709d1a2c7564f8ce6a9bb707e7f5451117ba80f8e8e2302662c29bc61246a6"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:b41c919caf1bb07fbacb0e29899ac513f9eb11cd4dd695975ad352c2c03cc4d7"
 
 [[skill]]
 name = "finishing-a-development-branch"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/finishing-a-development-branch"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:9edba9a38684c060fdc38290f640e1dc0c37de286723ac9be73bacacf7cd6f3d"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:932d23af11a3eb2b758f5b1d9aec7b6d28ddd26658328414f87f34dff7179d19"
 
 [[skill]]
 name = "ion-cli"
 source = ""
+kind = "git"
+commit = ""
 checksum = "sha256:bfc55c2c3f2bdc01ee30bd014d4b3f4bec6c60e6b2fab4d2526b7b893db5aa82"
 
 [[skill]]
 name = "receiving-code-review"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/receiving-code-review"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
 checksum = "sha256:2760c85d4f4117b0006e7ba755f4bbd61f8f4c185f347999763c97f507274e30"
 
 [[skill]]
 name = "requesting-code-review"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/requesting-code-review"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:046180bd8519589b176d575e32a720c43cf6834f2e4d182277faf8943fe03564"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:5a53d1f7f4cee4dee017a50d735335c7aae1874b2746207e076b5d9035315599"
 
 [[skill]]
 name = "subagent-driven-development"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/subagent-driven-development"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:96697eb8dde8631ee7cd5e1cb29c6d7c0b8ef2869c0875dd2eedf07762e5feae"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:22847d331db0aaedde0738d794896658020d25b703136d64f6690ee944688b7c"
 
 [[skill]]
 name = "systematic-debugging"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/systematic-debugging"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:6fddb1f3952f65f1f3a2ef267ae4ab6919ff70a8b36672ea908640ae21a34ea0"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:ebdc9e15e5745a87d6342b6ebbc90d1ac2766b95f2334a9450fc696842815e4d"
 
 [[skill]]
 name = "test-driven-development"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/test-driven-development"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
 checksum = "sha256:126f1ebf6ccd414f42544f6e83d8cc5adb089e1108eaffb7c400701e37eecd9f"
 
 [[skill]]
 name = "using-git-worktrees"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/using-git-worktrees"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:52bbb4b6e80918e83e92a1514f3b3757712154c2a8a42de24919e48a794c54fc"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:bd10bc57c23040630b17c877a9bc7b4be40c35161a2e2f110d770d78e3b0a4ff"
 
 [[skill]]
 name = "using-superpowers"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/using-superpowers"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:df560ba219fc828fd3e49749dffad1fc4d730538ea2ef81ec18751c5fb16da32"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:114207d2bda1336e283288aed358f51e1a78bbcd42e7d4e3e16a89095a89b006"
 
 [[skill]]
 name = "verification-before-completion"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/verification-before-completion"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
 checksum = "sha256:9b446f0c7fe1cfb560b1d34439523b1a76d5f177290007b2c053a1c749a4a8ba"
 
 [[skill]]
 name = "writing-plans"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/writing-plans"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:51c3a157b6026417aa8005230aafabc17bec820248c7c8b4ca106cf696c31ffd"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:6525448d840035c66700b81fbbfeb77431df04316615869569393f62af6a5616"
 
 [[skill]]
 name = "writing-skills"
 source = "https://github.com/obra/superpowers.git"
+kind = "git"
 path = "skills/writing-skills"
-commit = "363923f74aa9cd7b470c0aaa73dee629a8bfdc90"
-checksum = "sha256:5cb2cd1d6c3a2d4f598fdaceb852f302de034f57ee7872924ffc1fc2e9ca7745"
+commit = "6fd4507659784c351abbd2bc264c7162cfd386dc"
+checksum = "sha256:272ccd701f41b94e1119aaa99455a52cdb8e07283cb6a6aab32a7722d96bef45"
