+++ crates/bloqade-lanes-search/src/search/mod.rs
@@ -1,15 +1,12 @@
 //! High-level search facade.
 //!
-//! Slice-1 split of the pre-split `solve.rs` god module:
-//!
-//! - [`result`] — `SolveResult` / `SolveStatus` and their constructors.
+//! - [`result`] — `SolveResult` / `SolveStatus` / `CandidateAttempt` /
+//!   `MultiSolveResult` and their constructors.
 //! - [`options`] — `Strategy` / `InnerStrategy`, `SolveOptions` /
 //!   `EntropyOptions` / `EntanglingOptions`, with helpers
 //!   (`upgraded_for_entangling`, `clipped_future_layers`).
 //! - [`restarts`] — `run_with_components`, `pick_best`, `extract`
 //!   (strategy dispatch + restart orchestration).
-//! - [`solve`] — `CandidateAttempt` / `MultiSolveResult` result types plus
-//!   the option/result re-exports under the original `solve::*` path.
 //! - [`move_search`] — `MoveSearch` composition layer.
 //! - [`target_solver`] — `TargetSolver` (single-candidate solver wrapping
 //!   `SearchEngine` + `MoveSearch`).
@@ -19,5 +16,4 @@ pub mod move_search;
 pub mod options;
 pub mod restarts;
 pub mod result;
-pub mod solve;
 pub mod target_solver;
